test(sdk): one FakeMcp / FakeStore / recordingTool instead of 21
FakeMcp had 14 definitions across tests/ in 5 shapes that differed only in
whether setServers existed and in an error string nothing asserts; FakeStore
had 4 identical copies; recordingTool had 3. They now live in
tests/helpers/engine-fakes.ts, which already held FakeGate / FakeHookRunner and
was only imported by one suite.

One behavioural difference was real rather than cosmetic and is now explicit:
audit-t50-batch-i.test.ts's local recordingTool hardcoded `readOnly: true`
while every other copy defaulted to false, so its five call sites pass
{ readOnly: true }. Caught by K7 going red, not by inspection — the reason the
copies were checked one at a time rather than swapped wholesale.

Left alone deliberately: observability-v04.test.ts's FakeMcp reports real
statuses (not a fake-by-another-name), and the 16 makeConfig builders are
genuinely per-suite (15 distinct bodies).

Also drops the type imports the removals stranded, plus one that was already
unused (toolInputTruncationOf in t49-batch-b.test.ts).

3237 unit tests green.
