+++ projects/silver-core-sdk/tests/loop-control.test.ts
@@ -19,14 +19,11 @@ import {
 import { DefaultPermissionGate } from '../src/permissions/gate.js';
 import type {
   BuiltinTool,
-  CallToolResult,
   EngineConfig,
   EngineDeps,
-  McpRegistry,
 } from '../src/internal/contracts.js';
 import type {
   LoopStopProposal,
-  McpServerStatus,
   SDKMessage,
   SDKResultMessage,
   SDKSystemMessage,
@@ -38,25 +35,7 @@ import {
   toolUseReplyEvents,
 } from './helpers/mock-transport.js';
 import { makeSSEFetch } from './helpers/sse-fetch.js';
-
-class FakeMcp implements McpRegistry {
-  async connectAll(): Promise<void> {}
-  statuses(): McpServerStatus[] {
-    return [];
-  }
-  allTools(): [] {
-    return [];
-  }
-  has(_qualifiedName: string): boolean {
-    return false;
-  }
-  async call(): Promise<CallToolResult> {
-    return { content: [{ type: 'text', text: 'unexpected mcp call' }], isError: true };
-  }
-  async reconnect(_serverName: string): Promise<void> {}
-  setEnabled(_serverName: string, _enabled: boolean): void {}
-  async closeAll(): Promise<void> {}
-}
+import { FakeMcp } from './helpers/engine-fakes.js';
 
 const noHooks = {
   hasHooks: () => false,
