+++ projects/silver-core-sdk/tests/stop-hook-block.test.ts
@@ -18,37 +18,16 @@ import type {
   EngineConfig,
   EngineDeps,
   HookRunner,
-  McpRegistry,
 } from '../src/internal/contracts.js';
 import type {
   APIMessageParam,
-  CallToolResult,
   HookEvent,
   HookInput,
-  McpServerStatus,
   SDKMessage,
   SDKResultMessage,
 } from '../src/types.js';
 import { MockTransport, textReplyEvents } from './helpers/mock-transport.js';
-
-class FakeMcp implements McpRegistry {
-  async connectAll(): Promise<void> {}
-  statuses(): McpServerStatus[] {
-    return [];
-  }
-  allTools(): [] {
-    return [];
-  }
-  has(_qualifiedName: string): boolean {
-    return false;
-  }
-  async call(): Promise<CallToolResult> {
-    return { content: [{ type: 'text', text: 'unexpected mcp call' }], isError: true };
-  }
-  async reconnect(_serverName: string): Promise<void> {}
-  setEnabled(_serverName: string, _enabled: boolean): void {}
-  async closeAll(): Promise<void> {}
-}
+import { FakeMcp } from './helpers/engine-fakes.js';
 
 /** Scripted Stop-hook runner: one AggregatedHookResult per Stop invocation. */
 class ScriptedStopHooks implements HookRunner {
